Final solution. See comments.
The thing is that we can treat adding to everybody else
as removing chocolates from that exact intern.
So we need to count number of reduce operations for each intern.

Took me couple of days to solve.
Had to look into duscussions section.
Which I'm not particularly happy about,

Used paper on the last day to finally understand the answer.

Here is a great idea though: SOLVE IT ON PAPER FIRST.
And only then go to editor.

Solving in editor takes most of the time,
sometimes with random guesse which makes it even worse.

Full understanding first, then coding.